chore: move components/ to src/
