+++ .storybook/config.js
@@ -2,7 +2,6 @@ import { configure } from '@storybook/react';
 
 function loadStories() {
   require('../stories/index.js');
-  // You can require as many stories as you need.
 }
 
 configure(loadStories, module);
